requirements.txt
streamlit==1.28.0
pandas==2.0.0
requests==2.31.0
openai==1.3.0
python-dotenv==1.0.0
gspread==5.10.0
oauth2client==4.1.3
supabase==2.3.0
postgrest-py==0.13.0
